Stop using ChaiJS
ChaiJS makes it easy to introduce bugs in the tests, because it relies
on property getters with side effects. dirty-chai addresses this issue,
however, it is not actively maintained and regularily causes dependency
version conflicts.

So, to keep it simple `elm-app-boilerplate` now uses the built-in `assert`
node module but feel free to use any assertion library you want.